+++ locale_emoji_flutter/lib/locale_emoji_flutter.dart
@@ -1,5 +1,6 @@
 library locale_emoji_flutter;
 
+/// Export original package `locale_emoji` for compatibility
 export 'package:locale_emoji/locale_emoji.dart';
 
-export 'src/extension.dart';
+export 'src/extension.dart' show FlagEmojiLocaleExt;