+++ locale_emoji/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.3
+
+* Add doc comments.
+* Update package description.
+
 ## 0.1.2
 
 * Add separate package `locale_emoji_flutter` to ease use in Flutter apps.