+++ locale_emoji_flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.3
+
+* Add doc comments.
+* Add example.
+* Update package description.
+
 ## 0.1.1
 
 * Fix README link.